Crate chess_rs

A chess engine library written in rust

Modules

board

move

types

Structs

ChessGame

A state machine for a chess game

GameState

The state of a chess game in one specific position

PositionState

The state of a chess game in one specific position, excluding the piece positions/board and containing only metadata

Enums

GameTermination

A type describing all kinds of terminations of a chess game

MoveError

An error type for ChessGame::move.