chess_rs::move

Struct ExtendedMoveInfo

pub struct ExtendedMoveInfo {
    pub piece: ChessPiece,
    pub from: BoardPosition,
    pub to: BoardPosition,
    pub captured_piece: Option<(ChessPiece, BoardPosition)>,
    pub promotion: Option<PromotionPiece>,
    pub castling: Option<(BoardPosition, BoardPosition)>,
    pub en_passant: bool,
    pub disabled_castling_white: (bool, bool),
    pub disabled_castling_black: (bool, bool),
    pub set_en_passant_square: Option<BoardPosition>,
    pub check_event: Option<CheckEvent>,
}

A redundant QOL type useful for gui, for example to show popup screens for promotions, play visual or audio effects if some special moves occur or a check happens or whatever, like knowing what kind of piece got captured or something.

This is to spare the poor gui implementators the work of needing to derive these purely from diffing a chessboard

Fields
piece: ChessPiece

The piece that was moved

from: BoardPosition

From where the piece was moved

to: BoardPosition

Where the piece was moved to

captured_piece: Option<(ChessPiece, BoardPosition)>

If the move captures a piece, and if so, on which position that piece used to be on

promotion: Option<PromotionPiece>

If the move is a promotion, contains which piece the pawn promoted to. If implementing some kind of gui for this, noticing multiple different valid moves to the same square with different promotion options, this might be a good time to show a popup to let the user decide which piece they want to promote to.

castling: Option<(BoardPosition, BoardPosition)>

If castling occured, this contains the positions for the rook which moved with, more specifically Some((from, to))

en_passant: bool

If the move executed was en passant

disabled_castling_white: (bool, bool)

If the move disabled castling for white on (queen-side, king-side) respectively. This being false does not imply castling is enabled, just that no change in castling availability happened for this move.

disabled_castling_black: (bool, bool)

If the move disabled castling for black on (queen-side, king-side) respectively. This being false does not imply castling is enabled, just that no change in castling availability happened for this move.

set_en_passant_square: Option<BoardPosition>

If the move was a 2 step pawn move and enabled a potential en passant opportunity on the specified square behind it

check_event: Option<CheckEvent>

If the move was some kind of check event

Implementations

impl ExtendedMoveInfo

pub fn new(move_info: MoveInfo, check_event: Option<CheckEvent>) -> Self